move files into their respective packages
