+++ app/src/main/java/com/example/recipegenie/model/AppDatabase.kt
@@ -1,4 +1,4 @@
-package com.example.recipegenie
+package com.example.recipegenie.model
 
 
 import android.content.Context
