+++ app/src/main/java/com/example/recipegenie/view/RecipePage.kt
@@ -1,10 +1,15 @@
-package com.example.recipegenie
+package com.example.recipegenie.view
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.TextView
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import com.example.recipegenie.R
+import com.example.recipegenie.model.RecipeRepository
+import com.example.recipegenie.UpdateRecipe
+import com.example.recipegenie.model.Recipe
+import com.example.recipegenie.viewmodel.MainViewModel
 import com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
 
 class RecipePage : AppCompatActivity() {
