+++ app/src/main/java/com/example/recipegenie/view/SearchRecipes.kt
@@ -1,11 +1,12 @@
-package com.example.recipegenie
+package com.example.recipegenie.view
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.recipegenie.R
 
-class Landing_page : AppCompatActivity() {
+class SearchRecipes : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_landing_page)
+        setContentView(R.layout.activity_search_recipes)
     }
 }
