+++ app/src/main/java/com/example/recipegenie/view/UpdateRecipte.kt
@@ -5,6 +5,10 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.TextView
 import android.widget.Toast
+import com.example.recipegenie.model.Recipe
+import com.example.recipegenie.model.RecipeRepository
+import com.example.recipegenie.view.FavoritesList
+import com.example.recipegenie.viewmodel.MainViewModel
 import com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
 import com.google.android.material.textfield.TextInputEditText
