+++ app/src/main/java/com/example/recipegenie/viewmodel/RecipeAdapter.kt
@@ -1,10 +1,12 @@
-package com.example.recipegenie
+package com.example.recipegenie.viewmodel
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.example.recipegenie.R
+import com.example.recipegenie.model.Recipe
 
 class RecipeAdapter (private val onCardClick: (position: Int) -> Unit,
                     private val recipeList: List<Recipe>)
