+++ app/src/main/res/layout/activity_recipe_page.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/background_color"
     android:orientation="vertical"
-    tools:context=".RecipePage">
+    tools:context=".view.RecipePage">
 
     <ScrollView
         android:layout_width="match_parent"
